requirements.txt
fastapi==0.95.1
uvicorn==0.22.0

#new
SQLAlchemy==2.0.13
psycopg2-binary==2.9.6    #linux/mac user use: psycopg2-binary==2.9.6

python-dotenv==1.0.0
alembic==1.11.1
pydantic[email]==1.10.9
passlib[bcrypt]==1.7.4

asyncio

pytest==7.4.0
httpx==0.24.1

pre-commit==3.3.3
python-jose==3.3.0
python-multipart==0.0.6

Jinja2==3.1.2
aiofiles==23.1.0

python-socketio[asyncio_client]==5.10.0
jwt==1.3.1
pymongo==4.6.1
pyjwt==2.8.0
validators
requests
fastapi-mail==1.3.1
aiohttp

python-dateutil
numpy
faiss-cpu
groq
langchain
sentence_transformers
langchain-community